refactor(minifier): isolate compression pass finalization (#24651)
Normalize and recurring peephole passes share the same completion pipeline, but that pipeline lived inside the peephole optimizer. This obscured its ownership and made the required ordering between reference pruning, direct-eval refresh, and reachability analysis harder to follow.

Move pass completion into a dedicated private module while leaving mutation collection in the traversal context. This is a mechanical ownership change; convergence signals and operation ordering remain unchanged.

No measurable performance change is expected. Output is byte-identical to the base with identical iteration counts; CodSpeed CI is the performance evidence.

## Stack

1. **#24651 — pass finalization**
2. #24652 — pass transaction
3. #24653 — compression configuration

Verified with the full `oxc_minifier` test suite, clippy with warnings denied, formatting, minsize and allocation regeneration, and differential compress-only and full-minify corpora; snapshots are unchanged.

Implemented with AI assistance (OpenAI Codex). The contributor remains responsible for reviewing, understanding, and submitting the changes under the repository AI usage policy.
