+++ crates/oxc_minifier/src/lib.rs
@@ -44,6 +44,7 @@
 //!
 //! See the [crate documentation](https://github.com/oxc-project/oxc/tree/main/crates/oxc_minifier) for more details.
 
+mod compression_pass;
 mod compressor;
 pub(crate) mod generated;
 mod keep_var;
