+++ crates/oxc_minifier/src/peephole/remove_unused_private_members.rs
@@ -52,8 +52,8 @@ impl<'a> PeepholeOptimizations {
             if !keep {
                 // The element is being silently dropped from the vector. Walk
                 // its subtree so identifier references inside (e.g. a method
-                // body) are recorded in `PassDirty::dead_refs` and pruned by
-                // `flush_pass_dirty`. Without this, refs leak across passes
+                // body) are recorded in `PassChanges::removed_references` and pruned by
+                // `flush_pass_changes`. Without this, refs leak across passes
                 // and break idempotency. `drop_class_element` also records a
                 // mutation for the fixed-point loop driver.
                 ctx.drop_class_element(element);
