+++ crates/oxc_minifier/src/traverse_context/mod.rs
@@ -20,7 +20,7 @@ use crate::{
 
 mod ancestry;
 mod bound_identifier;
-mod drop_diff;
+mod dropped_subtree_collector;
 mod ecma_context;
 mod maybe_bound_identifier;
 mod reusable;
@@ -30,7 +30,7 @@ mod uid;
 use ancestry::PopToken;
 pub use ancestry::TraverseAncestry;
 pub use bound_identifier::BoundIdentifier;
-pub use drop_diff::as_direct_eval_call;
+pub use dropped_subtree_collector::as_direct_eval_call;
 pub use maybe_bound_identifier::MaybeBoundIdentifier;
 pub use reusable::ReusableTraverseCtx;
 pub use scoping::TraverseScoping;
