+++ crates/oxc_minifier/tests/peephole/dead_code_elimination.rs
@@ -234,7 +234,7 @@ fn dce_var_hoisting() {
 
 // Dropping a dead-after-throw statement (`module.exports = x`) removes the
 // only reference to `x`. Without recording that as a mutation, the peephole
-// loop terminates before `flush_pass_dirty` prunes the dropped reference,
+// loop terminates before `flush_pass_changes` prunes the dropped reference,
 // leaving the unused-declarator pass to see a stale reference and keep
 // `var x = {}`.
 #[test]
