+++ crates/oxc_minifier/tests/peephole/normalize.rs
@@ -35,7 +35,7 @@ fn test_void_ident() {
 }
 
 // Leak regression: Normalize runs before the peephole fixed-point loop, but
-// `PassDirty` is live from `MinifierState::new`, so Normalize's typed-helper
+// `PassChanges` is live from `MinifierState::new`, so Normalize's typed-helper
 // drops are recorded like any pass's and consumed by the pre-loop `end_pass`
 // flush. A leaked read makes `x` look referenced, blocking
 // unused-declaration removal.
